requirements.txt
pandas
scikit-learn
fastapi
uvicorn
joblib
mlflow
pyodbc
python-dotenv